Refactor UnicodePlots integration into extension
