+++ src/ReinforcementLearningCore/src/core/hooks.jl
@@ -10,7 +10,6 @@ export AbstractHook,
     DoEveryNSteps,
     DoOnExit
 
-using UnicodePlots: lineplot, lineplot!
 using Statistics: mean, std
 using CircularArrayBuffers: CircularVectorBuffer
 import ReinforcementLearningBase: RLBase
@@ -169,20 +168,6 @@ function Base.push!(hook::TotalRewardPerEpisode,
 )
     push!(hook.rewards, hook.reward)
     hook.reward = 0
-    return
-end
-
-function Base.show(io::IO, hook::TotalRewardPerEpisode{true, F}) where {F<:Number}
-    if length(hook.rewards) > 0
-        println(io, lineplot(
-            hook.rewards,
-            title="Total reward per episode",
-            xlabel="Episode",
-            ylabel="Score",
-        ))
-    else
-        println(io, typeof(hook))
-    end
     return
 end
